Updates version number with #defines, __stringify
One of the pains involved with developing out-of-core kernel
modules is not knowing which version of your module the user may
have loaded (as it's independent of the kernel version).  Adding
a /dev/cpu/msr_safe_version device helps, but it does not allow
for easy checking for changes in the API.

This commit adds #defines MSR_SAFE_MAJ_VER and MSR_SAFE_MIN_VER
and uses them (with __stringify()) to pass the appropriate
string to MODULE_VERSION().  Other code can now check the numeric
values of MSR_SAFE_MAJ_VER and MSR_SAFE_MIN_VER without having
to do string parsing in the kernel.