Adds new _ex structs for batchs arrays and ops.
This plays stupid struct layout games.  The first __u32 of
the incoming ioctl pointer will be interpreted as a pointer
to msr_batch_array if nonzero and a pointer to msr_batch_array_ex
if 0.  The latter will then compare the kernel module version
with the requested version and proceed only if they match.